fix(VoxelSize): `.io.utils` is not used by `.io` but a module for VoxelSize
