+++ plantseg/io/__init__.py
@@ -2,7 +2,7 @@ from plantseg.io.h5 import H5_EXTENSIONS, create_h5, load_h5, read_h5_voxel_size
 from plantseg.io.io import allowed_data_format, smart_load
 from plantseg.io.pil import PIL_EXTENSIONS, load_pil
 from plantseg.io.tiff import TIFF_EXTENSIONS, create_tiff, load_tiff, read_tiff_voxel_size
-from plantseg.io.utils import VoxelSize
+from plantseg.io.voxelsize import VoxelSize
 from plantseg.io.zarr import ZARR_EXTENSIONS, create_zarr, load_zarr, read_zarr_voxel_size
 
 # Use __all__ to let type checkers know what is part of the public API.
