+++ plantseg/io/zarr.py
@@ -13,7 +13,7 @@ from pathlib import Path
 import numpy as np
 import zarr
 
-from plantseg.io.utils import VoxelSize
+from plantseg.io.voxelsize import VoxelSize
 
 ZARR_EXTENSIONS = [".zarr"]
 ZARR_KEYS = ["raw", "predictions", "segmentation"]
